requirements.txt
streamlit
tensorflow
keras
numpy
pillow
matplotlib
scikit-learn
jupyter
python-telegram-bot
python-dotenv